Stack Implementation seperated into a diff. file
Seperated Stack implementation and main function
